+++ crates/xcomponent/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add frame-callback unregistration with Rust callback cleanup.
+
 ## [0.4.0](https://github.com/ohos-rs/ohos-native-bindings/compare/ohos-xcomponent-binding-v0.3.2...ohos-xcomponent-binding-v0.4.0) - 2026-08-26
 
 ### Added